css/cssom-view/elementsFromPoint-display-contents.html

<!DOCTYPE html>
<html>
<head>
    <title>CSSOM View - 5 - extensions to the Document interface</title>
    <link rel="author" title="Luke Warlow" href="mailto:lwarlow@igalia.com">
    <link rel="help" href="http://www.w3.org/TR/cssom-view/#extensions-to-the-document-interface">
    <meta name="flags" content="dom">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
        html,body {
            padding: 0;
            margin: 0;
        }
        #targetDiv {
            display:contents;
        }
    </style>
</head>
<body>
    <div id="targetDiv">Text Content</div>
    <script>
        var elements = document.elementsFromPoint(5, 5);

        test(() => {
            assert_equals(elements.length, 2);
            assert_equals(elements[0].tagName, "BODY");
            assert_equals(elements[1].tagName, "HTML");
        }, "elementsFromPoint should not include display:contents elements");
    </script>
</body>
</html>
